Add SanicBase
